Backend master omer productlist

+++ Projects/Backend/src/product/views.py
@@ -1,6 +1,7 @@
 from django.http import HttpResponse, Http404
 from rest_framework import status
 from rest_framework.authentication import TokenAuthentication
+from rest_framework.parsers import FileUploadParser
 from rest_framework.permissions import IsAuthenticated
 
 from rest_framework.response import Response
@@ -13,7 +14,7 @@ from product.models import Product, ProductList, SubOrder, Comment
 from product.serializers import ProductSerializer, ProductListSerializer, CommentSerializer
 
 # Create your views here.
-from user.models import User, Customer
+from user.models import User, Customer, Vendor
 from user.serializers import UserSerializer
 
 class ProductListAPIView(APIView):
@@ -24,11 +25,16 @@ class ProductListAPIView(APIView):
         return Response(serializer.data)
 
     def post(self, request):
-        serializer = ProductSerializer(data=request.data)
-        if serializer.is_valid():
+        if not Vendor.objects.filter(user_id=request.user.id).exists():
+            return Response({"message":"you must be logged in as a Vendor to add product"}, status=HTTP_400_BAD_REQUEST)
+        serializer = ProductSerializer(data=request.data, context={'request': request})
+        try:
             file = request.data['file']
             image = Product.objects.create(image=file)
-            serializer.save()
+        except:
+            None
+        if serializer.is_valid():
+            serializer.save(vendor_id=self.request.user.id)
             return Response(serializer.data, status=HTTP_201_CREATED)
         return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
 
@@ -58,9 +64,15 @@ class ProductDetailAPIView(APIView):
         return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
+        product = Product.objects.filter(id=id)
+        if not product.exists():
+            return Response({"message":"product not found"}, status=HTTP_404_NOT_FOUND)
+        product = product[0]
+        if not product.vendor_id == request.user.id:
+            return Response({"message":"you must be the owner to delete product"}, status=HTTP_400_BAD_REQUEST)
         product = self.get_product(id)
         product.delete()
-        return Response("product id "+ str(id) + " deleted", status=HTTP_204_NO_CONTENT)
+        return Response({"message":"product id "+ str(id) + " deleted"}, status=HTTP_204_NO_CONTENT)
 
 
 class ListListAPIView(APIView):
@@ -86,6 +98,16 @@ class ListListAPIView(APIView):
         product_lists = self.get_list_list(request, id)
         serializer = ProductListSerializer(product_lists, many=True, context={'request': request})
         return Response(serializer.data)
+
+    def post(self, request, id):
+        if not self.is_user(request, id):
+            return Response({"message": "bad request"}, status=status.HTTP_400_BAD_REQUEST)
+        # error handling done above
+        serializer = ProductListSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=HTTP_201_CREATED)
+        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
 
 class ListDetailAPIView(APIView):
 
@@ -121,18 +143,6 @@ class ListDetailAPIView(APIView):
         serializer = ProductListSerializer(list, context={'request': request})
         return Response(serializer.data)
 
-    def post(self, request, id, list_id):
-        if not self.check_user(request, id, list_id):
-            return Response({"message": "bad request"}, status=status.HTTP_400_BAD_REQUEST)
-        if not self.is_owner(request, id, list_id):
-            return Response({"message": "not allowed to access"}, status=status.HTTP_401_UNAUTHORIZED)
-        # error handling done above
-        serializer = ProductListSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=HTTP_201_CREATED)
-        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
-
     def put(self, request, id, list_id):
         if not self.check_user(request, id, list_id):
             return Response({"message": "bad request"}, status=status.HTTP_400_BAD_REQUEST)
@@ -140,11 +150,13 @@ class ListDetailAPIView(APIView):
         if not self.is_owner(request, id, list_id):
             return Response({"message": "not allowed to access"}, status=status.HTTP_401_UNAUTHORIZED)
         # error handling done above
-        serializer = ProductListSerializer(list, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
+        try:
+            list.name = request.data["name"]
+            list.save()
+            serializer = ProductListSerializer(list)
             return Response(serializer.data)
-        return Response(serializer.errors, status=HTTP_400_BAD_REQUEST)
+        except:
+            return Response({"message":"bad request"}, status=HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id, list_id):
         if not self.check_user(request, id, list_id):
